Conform to new test file structure
